requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
google-genai>=0.1.0
PyMuPDF>=1.22.0
openpyxl>=3.1.0
